+++ src/frontend/js/modules/vocabulary/index.ts
@@ -47,6 +47,5 @@ import './pages/word_list_app';
 import './pages/bulk_translate';
 import './pages/word_upload';
 import './pages/expression_interactable';
-import './pages/word_result_init';
 import './pages/translation_page';
 import './pages/starter_vocab';